Added ohlcBase to series for webgl
